# ccp/api/openapi/ccp-api.yaml
openapi: 3.1.0
info:
  title: ccp-api
  version: 2.0.0
  description: Authoritative Cloud Control Plane backend (Fork A1, ADR-0005). Actor is always the session; bodies never carry identity.
servers: [{ url: /v2 }]
security: [{ session: [] }]
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: ccp_session,
      description: "Opaque 256-bit token, HttpOnly Secure SameSite=Lax; server stores sha256(token). Non-GET also requires X-Ccp-Client header. The acting scope is X-Ccp-Project (a registered, ready project id); absent -> the reserved control-plane scope '@control' (data-birth spec: a header-less client is an inert CONTROL-PLANE client -- auth + admin-global + the projects registry only, never an implicit estate). '@control' is never listed by GET /projects, never a valid account binding, and every estate-only surface (request submission/approval, catalog reads) refuses it with 403 CONTROL_SCOPE." }
  parameters:
...
    id: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    E401: { description: No/expired/invalidated session, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } }
...
    E422: { description: Validation/manifest/policy failure, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } }
...
    E429: { description: Rate limited (Retry-After set), content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } }
  schemas:
    Err: { type: object, required: [code, reason], properties: { code: { type: string }, reason: { type: string }, details: {} } }
...
paths:
...
  /auth/reauth:
    post: { summary: 'ADR-0026 — re-authentication gate. EXACTLY ONE of {password} or {code} (a live TOTP code from any enrolled device; recovery codes never accepted). Success stamps reauthAt (ISO) on the CURRENT session item, valid REAUTH_MS=10 minutes; failures bump the same login lockout counter and a locked account is refused LOGIN_BACKOFF before the body is even checked',
      requestBody: { required: true, content: { application/json: { schema: { oneOf: [
        { type: object, required: [password], properties: { password: {type: string} }, additionalProperties: false },
        { type: object, required: [code], properties: { code: {type: string} }, additionalProperties: false } ] } } } },
      responses: { '200': { description: Elevated, content: { application/json: { schema: { type: object,
        required: [ok, reauthAt], properties: { ok: {type: boolean}, reauthAt: {type: string} } } } } },
        '401': { $ref: '#/components/responses/E401' }, '422': { $ref: '#/components/responses/E422' },
        '429': { $ref: '#/components/responses/E429' } } }
...
  /auth/totp-devices:
    get: { summary: 'List the caller''s enrolled TOTP devices (ADR-0024 §5) — id/name/enrolledAt/lastUsedAt only, NEVER secretEnc',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { type: object,
        required: [id, name, enrolledAt], properties: { id: {type: string}, name: {type: string}, enrolledAt: {type: string}, lastUsedAt: {type: string} } } } } } },
        '401': { $ref: '#/components/responses/E401' } } }
    post: { summary: 'Begin adding a device (⚿ re-auth required): mints a secret held on the CALLER''S FULL session (enrollSecretEnc/enrollOfferedAt), returns the QR/setup-key material. Offer expires after TOTP_PENDING_MS (5 minutes). Refused at the 5-device cap',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: object,
        required: [secret, otpauthUri], properties: { secret: {type: string}, otpauthUri: {type: string} } } } } },
        '401': { $ref: '#/components/responses/E401' }, '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } },
        '422': { description: DEVICE_LIMIT, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } } } }
  /auth/totp-devices/confirm:
    post: { summary: 'Confirm the held secret with a live code + a name (⚿ re-auth required; trimmed, single-line, 1-40 chars, no control chars). Appends the named device (materializes totpDevices, deletes the legacy totp field — ADR-0024 clause 2, idempotent). The account''s FIRST device auto-issues recovery codes (ADR-0025 clause 2), returned once in this response',
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [code, name],
        properties: { code: {type: string}, name: {type: string, minLength: 1, maxLength: 40} } } } } },
      responses: { '200': { description: Added, content: { application/json: { schema: { type: object,
        required: [id, name, enrolledAt], properties: { id: {type: string}, name: {type: string}, enrolledAt: {type: string},
        recoveryCodes: {type: array, items: {type: string}, description: 'present ONLY on the account''s first device'} } } } } },
        '401': { $ref: '#/components/responses/E401' }, '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } },
        '422': { $ref: '#/components/responses/E422' } } }
  /auth/totp-devices/{id}:
    delete: { summary: 'Remove a device by id (⚿ re-auth required). Refuses LAST_FACTOR if it is the last device AND needsTotp(account) is true. If the last device and removal IS allowed, the recovery-code set is deleted with it (audited both totp-device-remove and recovery-codes-clear)',
      parameters: [{ $ref: '#/components/parameters/id' }],
      responses: { '200': { description: Removed }, '401': { $ref: '#/components/responses/E401' },
        '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } },
        '404': { description: No such device }, '422': { description: LAST_FACTOR, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } } } }
  /auth/recovery-codes:
    get: { summary: 'Counts ONLY, ever — {remaining, generatedAt?}. Never the codes or their hashes, in any response',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: object, required: [remaining],
        properties: { remaining: {type: integer}, generatedAt: {type: string} } } } } },
        '401': { $ref: '#/components/responses/E401' } } }
  /auth/recovery-codes/regenerate:
    post: { summary: 'Replace the WHOLE recovery-code set (⚿ re-auth required) — old codes all die. Refused TOTP_REQUIRED when no device is enrolled (codes exist only while 2FA is active). Returns the 10 fresh plaintext codes EXACTLY ONCE',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: object, required: [codes, generatedAt],
        properties: { codes: {type: array, items: {type: string}, minItems: 10, maxItems: 10}, generatedAt: {type: string} } } } } },
        '401': { $ref: '#/components/responses/E401' }, '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } } } }
  /auth/sessions:
    get: { summary: 'The caller''s own LIVE sessions (expired/pre-session rows filtered out). id is the stored sha256 of the token (never the token itself); current marks the session resolved for THIS request',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { type: object,
        required: [id, issuedAt, lastSeenAt, current], properties: { id: {type: string}, issuedAt: {type: string}, lastSeenAt: {type: string}, current: {type: boolean} } } } } } },
        '401': { $ref: '#/components/responses/E401' } } }
  /auth/sessions/{id}:
    delete: { summary: 'Revoke ONE of the caller''s OWN sessions (⚿ re-auth required) — 404 on any id not in their own list (no cross-user probing). Deleting the CURRENT session IS sign-out',
      parameters: [{ $ref: '#/components/parameters/id' }],
      responses: { '200': { description: Revoked }, '401': { $ref: '#/components/responses/E401' },
        '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } },
        '404': { description: No such session } } }
  /auth/sessions/revoke-others:
    post: { summary: 'Sign out every OTHER session (⚿ re-auth required), keeping the caller''s own alive — deliberately NO sessionVersion bump (that would kill the keeper too)',
      responses: { '200': { description: OK, content: { application/json: { schema: { type: object,
        required: [ok, revoked], properties: { ok: {type: boolean}, revoked: {type: integer} } } } } },
        '401': { $ref: '#/components/responses/E401' }, '403': { description: REAUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } } } } }
...